+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - plugin support [ggandor/lightspeed.nvim](https://github.com/ggandor/lightspeed.nvim) #103
+- feat: use native autocmds on nvim >=0.7
 - feat: use typing (types.od.HighlightStyle) inside `theme.lua`
 - feat: add non-exists highlight from `overrides` option in dev mode (fixed #99)
 - enhance: bright & dark colors updated in terminal themes